Split HoverProvider and opcodes into their own files
